+++ osmos-nn/src/neuron.rs
@@ -1,10 +1,10 @@
 pub struct Neuron {
-    pub bias: f32,
-    pub weight_list: Vec<f32>,
+    pub bias: f64,
+    pub weight_list: Vec<f64>,
 }
 
 impl Neuron {
-    pub fn new(bias: f32, weight_list: &[f32]) -> Self {
+    pub fn new(bias: f64, weight_list: &[f64]) -> Self {
         Self {
             bias,
             weight_list: weight_list.to_vec(),
@@ -15,16 +15,16 @@ impl Neuron {
         let bias = rand::Rng::gen_range(rng, -1.0..=1.0);
         let weight_list = (0..weight_list_len)
             .map(|_| rand::Rng::gen_range(rng, -1.0..=1.0))
-            .collect::<Vec<f32>>();
+            .collect::<Vec<f64>>();
         Self { bias, weight_list }
     }
 
-    pub fn feed(&self, input_list: &[f32]) -> f32 {
+    pub fn feed(&self, input_list: &[f64]) -> f64 {
         let sum = input_list
             .iter()
             .zip(&self.weight_list)
             .map(|(input, weight)| input * weight)
-            .sum::<f32>();
+            .sum::<f64>();
         sum + self.bias
     }
 }
